requirements.txt
numpy==1.23.2
pandas==1.4.4
humanfriendly==10.0
scipy==1.9.1
python-dotenv==1.0.0
